Crate manifest

Manifest is a limit order book exchange on the Solana blockchain.

Modules

logs

program

quantities

state

utils

validation

Macros

global_seeds_with_bump

global_vault_seeds

global_vault_seeds_with_bump

market_vault_seeds

market_vault_seeds_with_bump

require

Constants

ID

The const program ID.

Statics

security_txt

Static string containing the security.txt file.

Functions

check_id

Returns true if given pubkey is the program ID.

entrypoint⚠

Safety

id

Returns the program ID.

process_instruction