manifest::state
Module market

Modules

market_helpers

Structs

AddOrderToMarketArgs

AddOrderToMarketResult

MarketFixed

MarketUnusedFreeListPadding

Functions

get_helper_ask_order

get_helper_bid_order

get_helper_order

get_helper_seat

Read a RBNode<ClaimedSeat> in an array of data at a given index.

get_mut_helper_ask_order

get_mut_helper_bid_order

get_mut_helper_order

get_mut_helper_seat

Read a RBNode<ClaimedSeat> in an array of data at a given index.

update_balance

Type Aliases

Bookside

BooksideReadOnly

ClaimedSeatTree

ClaimedSeatTreeReadOnly

MarketRef

Full market reference type.

MarketRefMut

Full market reference type.

MarketValue

Fully owned Market, used in clients that can copy.