Struct manifest::state::dynamic_account::DynamicAccount

pub struct DynamicAccount<Fixed, Dynamic> {
    pub fixed: Fixed,
    pub dynamic: Dynamic,
}

Fields
fixed: Fixeddynamic: Dynamic
Implementations

impl<Fixed: DerefOrBorrow<GlobalFixed>, Dynamic: DerefOrBorrow<[u8]>> DynamicAccount<Fixed, Dynamic>

pub fn get_balance_atoms(&self, trader: &Pubkey) -> GlobalAtoms

pub fn verify_min_balance(&self, trader: &Pubkey) -> ProgramResult

impl<Fixed: DerefOrBorrowMut<GlobalFixed>, Dynamic: DerefOrBorrowMut<[u8]>> DynamicAccount<Fixed, Dynamic>

pub fn global_expand(&mut self) -> ProgramResult

pub fn reduce( &mut self, trader: &Pubkey, num_atoms: GlobalAtoms ) -> ProgramResult

pub fn add_trader(&mut self, trader: &Pubkey) -> ProgramResult

Add GlobalTrader to the tree of global traders

pub fn evict_and_take_seat( &mut self, existing_trader: &Pubkey, new_trader: &Pubkey ) -> ProgramResult

Evict from the global account and steal their seat

pub fn add_order( &mut self, resting_order: &RestingOrder, global_trade_owner: &Pubkey ) -> ProgramResult

Add global order to the global account and specific market.

pub fn remove_order( &mut self, global_trade_owner: &Pubkey, global_trade_accounts: &GlobalTradeAccounts<'_, '_> ) -> ProgramResult

Remove global order. Update the GlobalTrader.

pub fn deposit_global( &mut self, trader: &Pubkey, num_atoms: GlobalAtoms ) -> ProgramResult

Deposit to global account.

pub fn withdraw_global( &mut self, trader: &Pubkey, num_atoms: GlobalAtoms ) -> ProgramResult

Withdraw from global account.

impl<Fixed: DerefOrBorrow<MarketFixed>, Dynamic: DerefOrBorrow<[u8]>> DynamicAccount<Fixed, Dynamic>

pub fn get_base_mint(&self) -> &Pubkey

pub fn get_quote_mint(&self) -> &Pubkey

pub fn impact_quote_atoms( &self, is_bid: bool, limit_base_atoms: BaseAtoms ) -> Result<QuoteAtoms, ProgramError>

pub fn impact_base_atoms( &self, is_bid: bool, round_up: bool, limit_quote_atoms: QuoteAtoms ) -> Result<BaseAtoms, ProgramError>

pub fn get_order_by_index(&self, index: DataIndex) -> &RestingOrder

pub fn get_trader_balance(&self, trader: &Pubkey) -> (BaseAtoms, QuoteAtoms)

pub fn get_trader_key_by_index(&self, index: DataIndex) -> &Pubkey

pub fn get_trader_voume(&self, trader: &Pubkey) -> QuoteAtoms

pub fn get_bids(&self) -> BooksideReadOnly<'_>

pub fn get_asks(&self) -> BooksideReadOnly<'_>

impl<Fixed: DerefOrBorrowMut<MarketFixed>, Dynamic: DerefOrBorrowMut<[u8]>> DynamicAccount<Fixed, Dynamic>

pub fn market_expand(&mut self) -> ProgramResult

pub fn claim_seat(&mut self, trader: &Pubkey) -> ProgramResult

pub fn get_trader_index(&mut self, trader: &Pubkey) -> DataIndex

pub fn release_seat(&mut self, trader: &Pubkey) -> ProgramResult

pub fn deposit( &mut self, trader: &Pubkey, amount_atoms: u64, is_base: bool ) -> ProgramResult

pub fn withdraw( &mut self, trader: &Pubkey, amount_atoms: u64, is_base: bool ) -> ProgramResult

pub fn place_order( &mut self, args: AddOrderToMarketArgs<'_, '_> ) -> Result<AddOrderToMarketResult, ProgramError>

Place an order and update the market

Check the order against the opposite bookside

Rest any amount of the order leftover on the book

pub fn cancel_order( &mut self, trader_index: DataIndex, order_sequence_number: u64, global_trade_accounts_opts: &[Option<GlobalTradeAccounts<'_, '_>>; 2] ) -> ProgramResult

pub fn cancel_order_by_index( &mut self, trader_index: DataIndex, order_index: DataIndex, global_trade_accounts_opts: &[Option<GlobalTradeAccounts<'_, '_>>; 2] ) -> ProgramResult